backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
sqlalchemy
psycopg2-binary
PyJWT
supabase
python-jose[cryptography]